Lucene issue on data indexing in "it-IT" culture
I found an issue in data indexing when current culture do not use DOT char as decimal separator (example: the decimal separator in Italian culture is Comma char).

I think that the issue is inside Lucene.NET. I do not found the bug, but I found a work around that works properly.

The solution is to invoke `StartIndexing` of "Examine.LuceneEngine.Providers.LuceneIndexer" class with Invariant Culture.  
The method is invoked in 2 line code. The first is to index data in syncronous mode, the second is in asyncronous mode.  

This patch solve for asyncronous mode.  
To solve the syncronous mode I suggest to create a new syncronous Task. But I have not many experience to it.